Move `MULTIPLE_VALUE` constant to constants folder
