+++ assets/src/edit-story/components/form/index.js
@@ -14,13 +14,8 @@
  * limitations under the License.
  */
 
-export {
-  default as MULTIPLE_VALUE,
-  MULTIPLE_DISPLAY_VALUE,
-} from './multipleValue';
 export { default as Button } from './button';
 export { default as Color } from './color';
-export { default as IconGroup } from './iconGroup';
 export { default as Input } from './input';
 export { default as Label } from './label';
 export { default as LinkInput } from './link';
@@ -28,7 +23,6 @@ export { default as Media } from './media';
 export { default as Numeric } from './numeric';
 export { default as BoxedNumeric } from './shared/boxedNumeric';
 export { default as Row } from './row';
-export { default as Spacer } from './spacer';
 export { default as Switch } from './switch';
 export { default as TextInput } from './text';
 export { default as TextArea } from './textArea';
