+++ assets/src/edit-story/components/form/toggleButton.js
@@ -27,7 +27,7 @@ import { useRef } from 'react';
  */
 import { KEYBOARD_USER_SELECTOR } from '../../utils/keyboardOnlyOutline';
 import { useKeyDownEffect } from '../../../design-system';
-import MULTIPLE_VALUE from './multipleValue';
+import { MULTIPLE_VALUE } from '../../constants';
 
 // Class should contain "mousetrap" to enable keyboard shortcuts on inputs.
 const CheckBoxInput = styled.input.attrs({
