+++ assets/src/edit-story/components/inspector/design/test/updateProperties.js
@@ -17,8 +17,8 @@
 /**
  * Internal dependencies
  */
+import { MULTIPLE_VALUE } from '../../../../constants';
 import updateProperties from '../updateProperties';
-import { MULTIPLE_VALUE } from '../../../form';
 
 describe('updateProperties', () => {
   let element;
