+++ assets/src/edit-story/components/panels/design/captions/captions.js
@@ -31,12 +31,14 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import {
+  MULTIPLE_VALUE,
+  MULTIPLE_DISPLAY_VALUE,
+} from '../../../../constants';
+import {
   Button,
   TextInput,
   Row,
   usePresubmitHandler,
-  MULTIPLE_VALUE,
-  MULTIPLE_DISPLAY_VALUE,
 } from '../../../form';
 import { useMediaPicker } from '../../../mediaPicker';
 import { SimplePanel } from '../../panel';
