+++ assets/src/edit-story/components/panels/design/link/link.js
@@ -36,7 +36,8 @@ import { useAPI } from '../../../../app/api';
 import { isValidUrl, toAbsoluteUrl, withProtocol } from '../../../../utils/url';
 import { Close } from '../../../../icons';
 import useElementsWithLinks from '../../../../utils/useElementsWithLinks';
-import { Media, Row, Button, LinkInput, MULTIPLE_VALUE } from '../../../form';
+import { MULTIPLE_VALUE } from '../../../../constants';
+import { Media, Row, Button, LinkInput } from '../../../form';
 import { useCanvas } from '../../../canvas';
 import { createLink } from '../../../elementLink';
 import { SimplePanel } from '../../panel';
