+++ assets/src/edit-story/components/panels/design/textStyle/test/textStyle.js
@@ -33,9 +33,12 @@ import AdvancedDropDown from '../../../../form/advancedDropDown';
 import ColorInput from '../../../../form/color/color';
 import createSolid from '../../../../../utils/createSolid';
 import CanvasContext from '../../../../canvas/context';
-import { MULTIPLE_VALUE, MULTIPLE_DISPLAY_VALUE } from '../../../../form';
+import {
+  HIDDEN_PADDING,
+  MULTIPLE_VALUE,
+  MULTIPLE_DISPLAY_VALUE,
+} from '../../../../../constants';
 import { renderPanel } from '../../../shared/test/_utils';
-import { HIDDEN_PADDING } from '../../../../../constants';
 
 jest.mock('../../../../../utils/textMeasurements');
 jest.mock('../../../../form/dropDown');
